Moved common_mysql package to server/platform/mysql
