+++ server/datastore/mysql/host_identity_scep.go
@@ -11,7 +11,7 @@ import (
 	"fmt"
 
 	"github.com/fleetdm/fleet/v4/ee/server/service/hostidentity/types"
-	"github.com/fleetdm/fleet/v4/server/datastore/mysql/common_mysql"
+	"github.com/fleetdm/fleet/v4/server/platform/mysql"
 	"github.com/jmoiron/sqlx"
 )
