+++ server/platform/mysql/batch.go
