+++ server/platform/mysql/batch_test.go
