+++ server/platform/mysql/errors.go
